# optimize watch-cache getlist