@sxllwx 2023-03-08
Update: 

```bash
# before
goos: linux
goarch: amd64
pkg: k8s.io/apiserver/pkg/storage/cacher
cpu: AMD EPYC 7K62 48-Core Processor
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=50-16         	      52	  21252051 ns/op	 1888993 B/op	    4921 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=50-16         	      54	  20719688 ns/op	 1848934 B/op	    4742 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=100-16        	      52	  21634518 ns/op	 3070568 B/op	    5370 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=100-16        	      54	  21208107 ns/op	 2987209 B/op	    5191 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=1000-16       	      51	  22207205 ns/op	 4495174 B/op	    5963 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=1000-16       	      52	  22786041 ns/op	 4425023 B/op	    5869 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=5000-16       	      40	  28037855 ns/op	15776801 B/op	   11318 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=5000-16       	      39	  31645133 ns/op	16156842 B/op	   11476 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=10000-16      	      30	  37621528 ns/op	31225948 B/op	   18409 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=10000-16      	      30	  35822193 ns/op	31225642 B/op	   18408 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=50000-16      	      12	 134383915 ns/op	146229170 B/op	   70962 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=50000-16      	      12	  96080972 ns/op	146228061 B/op	   70955 allocs/op

# after
goos: linux
goarch: amd64
pkg: k8s.io/apiserver/pkg/storage/cacher
cpu: AMD EPYC 7K62 48-Core Processor
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=50-16         	      49	  22527578 ns/op	 1875755 B/op	    5176 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=50-16         	      50	  22053190 ns/op	 1855078 B/op	    5074 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=100-16        	      51	  22239215 ns/op	 2075891 B/op	    4979 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=100-16        	      51	  22236347 ns/op	 2076071 B/op	    4979 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=1000-16       	      51	  22550354 ns/op	 2355291 B/op	    4979 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=1000-16       	      50	  22607157 ns/op	 2375407 B/op	    5078 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=5000-16       	      45	  25300896 ns/op	 4679899 B/op	    5639 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=5000-16       	      44	  24741316 ns/op	 4706111 B/op	    5768 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=10000-16      	      44	  26705112 ns/op	 7613999 B/op	    5769 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=10000-16      	      43	  26508674 ns/op	 7641156 B/op	    5902 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=50000-16      	      30	  34770585 ns/op	31674469 B/op	    8438 allocs/op
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=50000-16      	      31	  33819240 ns/op	31619322 B/op	    8167 allocs/op

# benchcmp

benchmark                                                                  old ns/op     new ns/op     delta
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=50-16        21252051      22527578      +6.00%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=50-16        20719688      22053190      +6.44%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=100-16       21634518      22239215      +2.80%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=100-16       21208107      22236347      +4.85%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=1000-16      22207205      22550354      +1.55%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=1000-16      22786041      22607157      -0.79%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=5000-16      28037855      25300896      -9.76%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=5000-16      31645133      24741316      -21.82%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=10000-16     37621528      26705112      -29.02%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=10000-16     35822193      26508674      -26.00%

benchmark                                                                  old allocs     new allocs     delta
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=50-16        4921           5176           +5.18%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=50-16        4742           5074           +7.00%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=100-16       5370           4979           -7.28%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=100-16       5191           4979           -4.08%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=1000-16      5963           4979           -16.50%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=1000-16      5869           5078           -13.48%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=5000-16      11318          5639           -50.18%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=5000-16      11476          5768           -49.74%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=10000-16     18409          5769           -68.66%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=10000-16     18408          5902           -67.94%

benchmark                                                                  old bytes     new bytes     delta
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=50-16        1888993       1875755       -0.70%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=50-16        1848934       1855078       +0.33%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=100-16       3070568       2075891       -32.39%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=100-16       2987209       2076071       -30.50%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=1000-16      4495174       2355291       -47.60%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=1000-16      4425023       2375407       -46.32%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=5000-16      15776801      4679899       -70.34%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=5000-16      16156842      4706111       -70.87%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=10000-16     31225948      7613999       -75.62%
BenchmarkCacher_GetList/totalObjectNum=50000,_expectObjectNum=10000-16     31225642      7641156       -75.53%
```